refactor: `BundleExtensions.kt` 파일 위치 변경
