+++ android/app/src/main/java/com/mulkkam/ui/settingwater/dialog/SettingWaterCupFragment.kt
@@ -10,7 +10,7 @@ import com.mulkkam.ui.binding.BindingBottomSheetDialogFragment
 import com.mulkkam.ui.settingwater.model.CupUiModel
 import com.mulkkam.ui.settingwater.model.CupUiModel.Companion.EMPTY_CUP_UI_MODEL
 import com.mulkkam.ui.settingwater.model.SettingWaterCupEditType
-import com.mulkkam.util.getParcelableCompat
+import com.mulkkam.util.extensions.getParcelableCompat
 
 class SettingWaterCupFragment :
     BindingBottomSheetDialogFragment<FragmentSettingWaterCupBinding>(
