+++ android/app/src/main/java/com/mulkkam/util/extensions/BundleExtensions.kt
@@ -1,4 +1,4 @@
-package com.mulkkam.util
+package com.mulkkam.util.extensions
 
 import android.os.Build
 import android.os.Bundle
